Add deserialization support for references, add a "zerocopy" example.
This exposes a soundness issue where a non-static input can be deserialized into a static reference...

Not sure what to do about it for now, will definitely have consequences upstream since this crate uses zero unsafe code to achieve this.

For now, a warning in the relevant example should be fine.