Mostly finish "Legacy" SNBT serializer, manually implement `Facet`
This'll be fun to merge with the finished work on my other machine :(

Facet's derive for `BorrowedRef` was not doing what I wanted, so it is now manually defined as an "iterator-like list".

The Facet derives for `RawTag` and `RawListTag` had to be removed since they were causing infinite recursion loops.

Oh and `no_std` is broken in `regex-lite`, so either a replacement needs to be found or regex-lite needs to be fixed...