chore: Mostly finalize serialized-size estimates
The unfinished part is that the sizes of structs and enums variants with fields are not yet calculated, leaving just about every non-builtin type with a hint of `TypeSerializeHint::None`.

Other than that I'm sure there are a few bugs and broken scenarios, but those are unavoidable at the moment as the current number of possible test cases is very small.