chore: Bump `facet` version, replace iterator with non-describing impls
Turns out facet-format already has a mechanism for deserializing types that are not self-describing.

Like other binary formats (bincode, postcard, etc), our types themselves are the data format that we are reading.
The bytes coming in can tell us about what we are reading, like string/vec/map lengths, but not what the actual type is.